+++ app/repository/get_file_url.tsx
@@ -1,0 +1,17 @@
+import { storage, ref, getDownloadURL } from "@/firebase";
+
+// Firebase Storageにファイルをアップロードする
+export const GetFileURL = async (filePath: string): Promise<string> => {
+  const fileRef = ref(storage, filePath);
+  const url = await getDownloadURL(fileRef)
+    .then((url) => {
+      console.log("successfully get file url", url);
+      return url;
+    })
+    .catch((error) => {
+      console.error("get file url was failed", error);
+      return "";
+    });
+
+  return url;
+};
