+++ app/repository/index.ts
@@ -1,1 +1,3 @@
 export * from "./post";
+export * from "./upload_file";
+export * from "./get_file_url";
